Add files via upload
Here we use HTML, CSS, JavaScript(mini) where we make Home page, Login page, Job Seekers, Employers, About, Contact page. In this website we login in this site and we upload the  details and the past of the details also need i.e., work experience and so more. According to past experience and past job we suggest the job which job is suitable for them and we upload resume and apply on job according to the available job. Here we provide Full Time, Part Time job available on this site. My logo is "Find the perfect job that you deserve.